+++ src/tbox/platform/poller.c
@@ -42,8 +42,8 @@
 #   ifdef TB_CONFIG_MICRO_ENABLE
 #       include "posix/poller_select.c"
 #   else
-#       include "posix/poller_select.c"
-//#       include "windows/poller_iocp.c"
+//#       include "posix/poller_select.c"
+#       include "windows/poller_iocp.c"
 #   endif
 #elif defined(TB_CONFIG_POSIX_HAVE_EPOLL_CREATE) \
     && defined(TB_CONFIG_POSIX_HAVE_EPOLL_WAIT)