+++ src/tbox/platform/windows/thread_local.c
@@ -99,6 +99,10 @@ tb_void_t tb_thread_local_exit(tb_thread_local_ref_t local)
     // check
     tb_assert(local);
 
+    // free the previous data first
+    if (local->free && tb_thread_local_has(local))
+        local->free(tb_thread_local_get(local));
+
     // exit it
     TlsFree(*((DWORD*)local->priv));
 }