+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * Fix setenv empty value crash 
 * Fix coroutine.sleep bug
 * Fix windows root path bug 
+* Fix thread local memory leak
 
 ## v1.6.2
 
@@ -166,6 +167,7 @@
 * 修复设置环境变量值为空时的崩溃问题
 * 修复协程sleep超时覆写数据的bug
 * 修复windows根路径问题
+* 修复tls线程存储内存泄露问题
 
 ## v1.6.2